board: kent: gfxboard changes (2/3)
This is patch two of three creating the kent_gfxboard u-boot port.

This patch shows the modifications that were made to the Atmel
sama5d27_som1_ek port to represent the hardware present on the Kent
board. These files were copied from the corresponding sama5d27_som1_ek
files in the previous patch.

The functional changes are:

- dts:
    - remove PDA LCD connections and 1-wire ID EEPROM
    - remove USB power control
    - Use 4-wire bus width instead of 8 for sdmmc0
    - remove sdmmc1
- board C file:
    - remove USB power control
    - change pin assignments for LEDs
- defconfig:
    - SUPPORT_RAW_INITRD=y to load bare cpio archive generated by yocto